+++ tests/Test.hs
@@ -1,25 +1,24 @@
 {-# LANGUAGE CPP #-}
 module Main (main) where
 
-import Test.Framework (defaultMain, testGroup, Test)
-import Test.Framework.Providers.HUnit
+import Test.Tasty (defaultMain, testGroup, TestTree)
 
 import qualified Prop.Quaternion
 import qualified Unit.Binary
 import qualified Unit.Plucker
 import qualified Unit.V
 
-tests :: [Test]
+tests :: [TestTree]
 tests =
   [ testGroup "Property tests"
     [ testGroup "Quaternion" Prop.Quaternion.tests
     ]
   , testGroup "Unit tests"
-    [ testGroup "Binary" $ hUnitTestToTests Unit.Binary.tests
-    , testGroup "Plucker" $ hUnitTestToTests Unit.Plucker.tests
-    , testGroup "V" $ hUnitTestToTests Unit.V.tests
+    [ testGroup "Binary" Unit.Binary.tests
+    , testGroup "Plucker" Unit.Plucker.tests
+    , testGroup "V" Unit.V.tests
     ]
   ]
 
 main :: IO ()
-main = defaultMain tests
+main = defaultMain $ testGroup "linear" tests